requirements.txt
Django>=4.0,<5.0
python-telegram-bot>=20.0
django-environ
environs
pysqlite3  
